Added number and channel check

+++ src/interactions/slashCommands/create.ts
@@ -9,7 +9,7 @@ export default {
 	data: new SlashCommandBuilder()
 		.setName(commandID)
 		.setDescription('create'),
-	async execute(_client: Client, interaction: ChatInputCommandInteraction) {
+	async execute(client: Client, interaction: ChatInputCommandInteraction) {
 		try {
 			if (!interaction.channel) return;
 
@@ -36,25 +36,36 @@ export default {
 			});
 
 			const filter = (m: Message) => m.author.id === interaction.user.id;
-			await interaction.reply({ embeds: [createEmbed] });
+			await interaction.reply({embeds: [createEmbed]});
 
-			const channel = await collectMessage(interaction.channel, filter, 3000);
+
+			const channel = await collectMessage(interaction.channel, filter, 60000);
 			if (!channel) return interaction.channel.send('test');
-			// TODO: Check if channel id valid
+
+			const channelid = channel.mentions.channels.first()?.id;
+			if (!channelid) return interaction.channel.send('VALIDE');
+
+			const dcChannel = await client.channels.fetch(channelid);
+			if (!dcChannel) return  interaction.channel.send('VALIDE2');
+
 			await channel.delete();
 			await editEmbed(createEmbed, channel.content, 'Pending', 'Pending', 'Please provide the range within which the game will be played. The range can be between 1 and 1,000,000');
 			await interaction.editReply({embeds: [createEmbed]});
 
-			const number = await collectMessage(interaction.channel, filter, 3000);
+
+			const number = await collectMessage(interaction.channel, filter, 60000);
 			if (!number) return interaction.channel.send('test2');
-			// TODO: Check if is it a positive number without decimal
+
+			const numberCheck = isValidNumber(number.content);
+			if (!numberCheck) return interaction.channel.send('VALIDEEE2');
+
 			await number.delete();
 			await editEmbed(createEmbed, channel.content, number.content, 'Pending', 'Please indicate the price for the game');
 			await interaction.editReply({embeds: [createEmbed]});
 
-			const price = await collectMessage(interaction.channel, filter, 3000);
+
+			const price = await collectMessage(interaction.channel, filter, 60000);
 			if (!price) return interaction.channel.send('test3');
-			// TODO: idk :/
 			await price.delete();
 			await editEmbed(createEmbed, channel.content, number.content, price.content, 'Thx');
 			await interaction.editReply({embeds: [createEmbed]});
@@ -101,4 +112,16 @@ async function editEmbed(embed: EmbedBuilder, v1:string, v2:string, v3:string, d
 		}
 	]);
 	embed.setDescription(description);
+}
+
+function isValidNumber(input: string) {
+	const number = parseInt(input, 10);
+
+	if (isNaN(number)) return false;
+
+	if (number < 1 || number > 1000000) return false;
+
+	if (number.toString() !== input) return false;
+
+	return true;
 }
